Fix several warnings
m_bmqstoragetool
================

FileManager
- Follow ISO C++ destructor rules

FileManagerMock
- Fix weak-vtable

SearchShortResult
- Remove unused d_allocator_p parameter
- Add allocator() method to retrieve allocator from subobject

bmq
===

bmqa_mocksession
- Remove unused const overload of uriCorrIdToQueues

bmqma_countingallocator
- Properly mark CountingAllocator::configureStatContextTableInfoProvider
as deprecated

bmqtst_logcontroller
- Add missing cases in switch statement

mqb
===

mqba_commandrouter
- Mark CommandRouter::ClusterWideRoutingMode::getRouteTargets
errorDescription as unused
- Move RoutingMode::~RoutingMode() to implementation file

mqba_sessionnegotiator
- Escape '@' character used by doxygen special command parsing

mqbcfg_tcpinterfaceconfigvalidator
- Add newline to end of header

mqbi_queueengine
- Mark QueueEngine::logAppSubscription appId as unused

mqbmock_storagemanager
- Force assertion failure for unimplemented StorageManager::fileStore

mqbnet_tcpsessionfactory
- Remove unused channelCast() helper
