+++ src/applications/bmqstoragetool/m_bmqstoragetool_filemanager.h
@@ -127,7 +127,7 @@ inline FileManagerImpl::FileHandler<ITER>::FileHandler(
 }
 
 template <typename ITER>
-inline FileManagerImpl::FileHandler<ITER>::~FileHandler()
+inline FileManagerImpl::FileHandler<ITER>::~FileHandler<ITER>()
 {
     d_iter.clear();
     if (d_mfd.isValid()) {
