+++ src/applications/bmqstoragetool/m_bmqstoragetool_searchresult.cpp
@@ -311,9 +311,13 @@ SearchShortResult::SearchShortResult(
 , d_printedMessagesCount(0)
 , d_guidMap(allocator)
 , d_guidList(allocator)
-, d_allocator_p(allocator)
 {
     // NOTHING
+}
+
+bslma::Allocator* SearchShortResult::allocator() const
+{
+    return d_guidMap.get_allocator().mechanism();
 }
 
 bool SearchShortResult::processMessageRecord(
