+++ src/groups/bmq/bmqma/bmqma_countingallocator.h
@@ -78,6 +78,7 @@
 #include <bslmf_nestedtraitdeclaration.h>
 #include <bslmt_mutex.h>
 #include <bsls_atomic.h>
+#include <bsls_deprecatefeature.h>
 #include <bsls_keyword.h>
 #include <bsls_types.h>
 
@@ -167,8 +168,10 @@ class CountingAllocator BSLS_KEYWORD_FINAL : public bslma::Allocator {
 
     /// Configure the specified `tableInfoProvider` with a filter, a
     /// comparator, and columns to print the statistics generated by this
-    /// `bmqma::CountingAllocator`.  @DEPRECATED: use the version taking a
-    /// `bmqst::Table` and `bmqst::BasicTableInfoProvider` instead.
+    /// `bmqma::CountingAllocator`.
+    BSLS_DEPRECATE_FEATURE("bmq",
+                           "configureStatContextTableInfoProvider",
+                           "Use the alternate overload instead.")
     static void configureStatContextTableInfoProvider(
         bmqst::StatContextTableInfoProvider* tableInfoProvider);
